Using static types for the sake of clarity.
Memory Pools can actually return any type depending on the type you pass
on, but I think I prefer having static guarantees in this case to avoid
making mistakes in the future.

Also, since we know that if the MemoryPool is not null, calling the
funciton is safe-ish so we remove the unsafety from that place.