Updated for the latest TW3 patch.
This makes effectively the tool not compatible with the older version.

MemoryPool was inlined, but then I noticed, that the function number 25
of the vtable of the memory pool object *is* the one that builds the
objects, so we can use that now.

Changed all the patterns since the code in the game itself changed a
lot.