Changed InputFloat for Drag for Position
As requested for our lord and savior, Jim2Point0, The Legendary
Fortographer, it would be more convenient for more the position to be
draggables.

Despite the current despair of our society, screenshots should be a top
priority of every peasant in this world, and so this commit had to be
made.

This is god's law.